Implement installer out of command
